# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Service Ticket Management System API"
  description: "API to manage a Mechanic Service Ticket System"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# Security Definitions
# =========================
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# =========================
# Paths
# =========================
paths: 
...
  /part_descriptions: #Parts Description CRUD
    post: #Create
      tags:
        - Part Descriptions
      summary: Create a new part description
      description: This endpoint allows you to create a new part description.
      parameters:
        - in: body
          name: body
          description: Part description object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/PartDescriptionPayload'
      responses:
        201:
          description: Part description created successfully
          schema:
            $ref: '#/definitions/PartDescriptionMessageResponse'
          examples:
            application/json:
              {
                "status": "success",
                "message": "Successfully created part description",
                "part_description": {
                  "id": 1,
                  "name": "Brake Pad",
                  "brand": "Brand A",
                  "price": 100.00
                }
              }
    get: #Read
      tags:
        - Part Descriptions
      summary: Get part descriptions (with pagination)
      description: This endpoint returns a list of part descriptions with pagination. 
        **You can use the `page` and `per_page` query parameters to control pagination.**
      parameters:
        - in: query
          name: page
          type: integer
          description: Page number for pagination.
        - in: query
          name: per_page
          type: integer
          description: Number of part descriptions per page.
      responses:
        200:
          description: Return all part descriptions
          schema:
            $ref: '#/definitions/PartDescriptionListResponse'
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "name": "Brake Pad",
                  "brand": "Brand A",
                  "price": 100.00
                }
              ]
  
  # --------- Parts Description Get By ID ---------      
  /part_descriptions/{id}:
    get: #Read
      tags:
        - Part Descriptions
      summary: Get part description by ID
      description: Retrieve a part description's details by its unique ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
          description: The ID of the item to retrieve.
      responses:
        200:
          description: Part description found
          schema:
            $ref: '#/definitions/PartDescriptionResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "Brake Pad",
                "brand": "Brand A",
                "price": 100.00
              }
        404:
          description: Part description not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Part description not found.
              status: error        
    put: #Update
      tags:
        - Part Descriptions
      summary: Update a part description
      description: Update an existing part description's details.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
          description: The ID of the item to retrieve.
        - in: body
          name: body
          description: Information used to update a part description
          required: true
          schema:
            $ref: '#/definitions/PartDescriptionPayload'
      responses:
        200:
          description: Part description updated successfully
          schema:
            $ref: '#/definitions/PartDescriptionMessageResponse'
          examples:
            application/json:
              {
                "status": "success",
                "message": "Successfully updated part description",
                "part_description": {
                  "id": 1,
                  "name": "Brake Pad",
                  "brand": "Brand A",
                  "price": 100.00
                }
              }
        404:
          description: Invalid part description ID
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Part description not found.
              status: error
    delete: #Delete
      tags:
        - Part Descriptions
      summary: Delete a part description
      description: This endpoint allows you to delete a part description.
      parameters:
        - in: path
          name: part_description_id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the item to retrieve.
      responses:
        200:
          description: Part description deleted successfully
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Successfully deleted part description 1.
              status: success
        404:
          description: Invalid part description ID
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Part description not found.
              status: error
  
  # --------- Parts Description Search ---------     
  /part_descriptions/search:
    get:
      tags:
        - Part Descriptions
      summary: Search part descriptions
      description: This endpoint allows you to search for part descriptions by name and/or brand.
        **At least one of `name` or `brand` must be provided as a query parameter.**.
      parameters:
        - in: query
          name: name
          type: string
          description: Search term to filter part descriptions by name.
        - in: query
          name: brand
          type: string
          description: Search term to filter part descriptions by brand.
      responses:
        200:
          description: Return part descriptions matching the search term
          schema:
            $ref: '#/definitions/PartDescriptionListResponse'
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "name": "Brake Pad",
                  "brand": "Brand A",
                  "price": 100.00
                }
              ]
        400:
          description: No query parameters provided
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: At least one search parameter (name or brand) is required.
              status: error
        404:
          description: No part descriptions found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: No part descriptions found.
              status: error

  # --------- Serialized Parts Endpoints ---------
...
definitions: 
...
  MessageSchemaResponse:
    type: object
    properties:
      message:
        type: string
      status:
        type: string

  # --------- Mechanic Schemas ---------
...
  PartDescriptionPayload:
    type: object
    properties:
      name:
        type: string
      brand:
        type: string
      price:
        type: integer
  PartDescriptionResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      brand:
        type: string
      price:
        type: integer
  PartDescriptionListResponse:
    type: array
    items:
      $ref: '#/definitions/PartDescriptionResponse'
  PartDescriptionMessageResponse:
    type: object
    properties:
      status:
        type: string
      message:
        type: string
      part_description:
        $ref: '#/definitions/PartDescriptionResponse'

#--------------- Serialized Part Schemas-------
...